Add AI chat system and float growth amounts
- Add AiChatServer.cs: embedded HTTP server with /api/grow, /api/proxy/chat,
  /api/characters, /api/config endpoints; growth jobs processed on Unity main
  thread; UnityWebRequest coroutine for TLS-capable HTTPS proxy
- Add AiChatWebUi.cs: self-contained browser UI for chatting with in-scene
  characters; AI triggers growth/shrink via hidden <growth> tags; supports
  "matthew says" override commands; displays growth in feet (0.35 units = 2 ft)
- Change Week and InflationAmount from int to float for sub-unit precision;
  AI amounts range from 0.05 to 1.0 per message
- Add SetInflationDirect() to reset _inflationChange alongside InflationAmount,
  fixing IsInflated staying false after shrink commands
- Add AI config entries (API key, URL, model, port, hotkey) to GrowthPlugin
- Update Timeline interpolable generic type to float
- Update README with AI chat setup and new configuration options

+++ PregnancyData.cs
@@ -128,7 +128,7 @@ namespace KK_Growth
 		public MenstruationSchedule MenstruationSchedule = MenstruationSchedule.Default;
 
 		// Token: 0x04000016 RID: 22
-		public int Week;
+		public float Week;
 
 		// Token: 0x04000017 RID: 23
 		public int PregnancyCount;
